pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "logtail-python>=0.3.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "isort>=6.0.1",
    "mypy>=1.17.1",
]
